requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pypdf2==3.0.1
python-multipart==0.0.6
faker==19.13.0
sqlite3== # Included in Python standard library
pydantic==2.6.1
pydantic-settings==2.1.0